initial stab at packaging functions and their dependencies as a docker image
Major changes:
- removed all packaging/zip/upload functionality from `tasks.py`
- consolidated functions and requirements
- introduced a Dockerfile for packaging into a docker image
- updated the CDK lambda function construct to build and use an image
  asset
- removed codebuild components
