+++ .gitignore
@@ -13,3 +13,4 @@ outfile.txt
 .cdk.staging
 cdk.out
 cdk.context.json
+node_modules
