--- buildspec.yml
@@ -1,22 +1,0 @@
-version: 0.2
-
-phases:
-  install:
-    runtime-versions:
-      python: 3.12
-  pre_build:
-    commands:
-      - python -m virtualenv venv
-      - source venv/bin/activate
-      - pip install -U pip pip-tools
-      - pip-sync requirements/dev.txt
-      - invoke test
-      - pip-sync requirements/base.txt
-  build:
-    commands:
-      - invoke deploy.all
-      - invoke release --description "$CODEBUILD_SOURCE_VERSION"
-artifacts:
-  files:
-    - dist/*.zip
-  discard-paths: yes
