+++ example.env
@@ -9,9 +9,6 @@ STACK_TAGS=Key=project,Value=MH Key=department,Value=DE
 # name of the cfn stack; also used to prefix many resource names (required)
 STACK_NAME=
 
-# s3 bucket to store packaged lambda code
-LAMBDA_CODE_BUCKET=
-
 # where cloudwatch alarm notifications will get sent
 NOTIFICATION_EMAIL=
