--- function_requirements/common-requirements.in
@@ -1,12 +1,0 @@
-boto3
-aws-lambda-logging
-requests
-tenacity
-PyJWT
-pytz
-python-dotenv
-google-api-python-client
-google-auth
-google-auth-oauthlib
-# constrain rsa version to something supported by awscli==1.27.142 in requirements/base.txt
-rsa==4.7.2
