--- function_requirements/common-requirements.txt
@@ -1,95 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.12
-# by the following command:
-#
-#    pip-compile common-requirements.in
-#
-aws-lambda-logging==0.1.1
-    # via -r common-requirements.in
-boto3==1.38.16
-    # via -r common-requirements.in
-botocore==1.38.16
-    # via
-    #   boto3
-    #   s3transfer
-cachetools==5.5.2
-    # via google-auth
-certifi==2025.4.26
-    # via requests
-charset-normalizer==3.4.2
-    # via requests
-google-api-core==2.24.2
-    # via google-api-python-client
-google-api-python-client==2.169.0
-    # via -r common-requirements.in
-google-auth==2.40.1
-    # via
-    #   -r common-requirements.in
-    #   google-api-core
-    #   google-api-python-client
-    #   google-auth-httplib2
-    #   google-auth-oauthlib
-google-auth-httplib2==0.2.0
-    # via google-api-python-client
-google-auth-oauthlib==1.2.2
-    # via -r common-requirements.in
-googleapis-common-protos==1.70.0
-    # via google-api-core
-httplib2==0.22.0
-    # via
-    #   google-api-python-client
-    #   google-auth-httplib2
-idna==3.10
-    # via requests
-jmespath==1.0.1
-    # via
-    #   boto3
-    #   botocore
-oauthlib==3.2.2
-    # via requests-oauthlib
-proto-plus==1.26.1
-    # via google-api-core
-protobuf==6.31.0
-    # via
-    #   google-api-core
-    #   googleapis-common-protos
-    #   proto-plus
-pyasn1==0.6.1
-    # via
-    #   pyasn1-modules
-    #   rsa
-pyasn1-modules==0.4.2
-    # via google-auth
-pyjwt==2.10.1
-    # via -r common-requirements.in
-pyparsing==3.2.3
-    # via httplib2
-python-dateutil==2.9.0.post0
-    # via botocore
-python-dotenv==1.1.0
-    # via -r common-requirements.in
-pytz==2025.2
-    # via -r common-requirements.in
-requests==2.32.3
-    # via
-    #   -r common-requirements.in
-    #   google-api-core
-    #   requests-oauthlib
-requests-oauthlib==2.0.0
-    # via google-auth-oauthlib
-rsa==4.7.2
-    # via
-    #   -r common-requirements.in
-    #   google-auth
-s3transfer==0.12.0
-    # via boto3
-six==1.17.0
-    # via python-dateutil
-tenacity==9.1.2
-    # via -r common-requirements.in
-uritemplate==4.1.1
-    # via google-api-python-client
-urllib3==2.4.0
-    # via
-    #   botocore
-    #   requests
