--- function_requirements/schedule-update.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
