--- function_requirements/slack.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
