--- function_requirements/status-query.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
