--- function_requirements/zoom-downloader.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
