--- function_requirements/zoom-log-notifications.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
