--- function_requirements/zoom-on-demand.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
