--- function_requirements/zoom-uploader.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
