--- function_requirements/zoom-webhook.txt
@@ -1,1 +1,0 @@
--r common-requirements.txt
