+++ package.json
@@ -1,5 +1,8 @@
 {
   "dependencies": {
     "aws-cdk": "^2.1014.0"
+  },
+  "scripts": {
+    "cdk": "cdk"
   }
 }
